fixed FPFRAME warnings for powerpc
2005-08-10  Andreas Tobler  <a.tobler@schweiz.ch>

        * config/powerpc/ppc_stack_frame.h (FPFRAME): Return an address and
        not a pointer.

+++ config/powerpc/ppc_stack_frame.h
@@ -83,8 +83,8 @@ typedef struct _exceptionFrame {
  * a NULL here otherwise we'll segfault when examining the top frame.
  */
 #define PCFRAME(f) ((f)->sp == 0 ? 0 : (f)->sp->lr - 1)
-/* Get the frame pointer from an exceptionFrame. */
-#define FPFRAME(f) ((f)->sp)
+/* Get the frame pointer address from an exceptionFrame. */
+#define FPFRAME(f) ((unsigned long)(f)->sp)
 
 /* Get the "this" object out of a stack frame. */
 #define FRAMEOBJECT(obj, f, einfo) \
